feat: client now sleeps when limits are hit.
Added response handlers for abuse and rate limit exceeded errors which
cause the app to sleep for the amount of time in the `retry-after`
response header or until `x-ratelimit-reset` respectively. Nyc is now
capable of being left running indefinitely. If uninterrupted it will
eventually scrape anything.

Still does not daemonise or restart if interupted.