feat: separated commits from commit comments.
Commit comments are a special GitHub feature and should always be
queried, whereas the commit log can be grabbed more efficiently in batch
operations by linking to a clone of the repo. The list of commit
comments is also in general much shorter than the commit log as well, so
query cost shouldn't be hit too hard. There is a cost, but this info is
important and there's nowhere to get it but the API.