Traceability Enhancements (#61)
* docs: updated CLI help message.

* refactor: changed argument passing from CLI.

We're getting more and more options, and we need a cleaner way to deal
with that.

* feat: made reactions and commits optional

Added individual flags to toggle them.

* fix: added a throttle to the request pipe.

We should no longer be hitting abuse alarms.

Closes #47

* fix: broken flags interaction.

commits and local-dir can't both be active.

* fix: set concurrency to 1.

When used from a fast internet connection, parallel requests set off
the abuse alarm. If the internet is slow we can do parallel requests,
but that's not something to optimise for just yet.

* fix: PullRequestComment doesn't exist.

Comments on PullRequests are IssueComments.

* Refactor: use a queue instead of a timeouts.

Much more transparent. Now we can see how many requests are queued up
and react to it.

* fix: Only grabbing commit logs from master.

This puts the GH API version inline with the planned local version.

* refactor: more sophisticated graphql, DSL.

* refactor: reformatted queries to new DSL.

* refactor: cleaner logging.

* refactor: Passing response metadata through pipe.

Better error handling and debugging.

* refactor: removed depagination boilerplate

It's now automatically provided in the query engine.

* feat: added --full option for uncompressed output

Returns the query response JSON with all of the internal logic cleaned
out and edges collapsed into lists.

* fix: reviews and reactions are working again.

* feat: grab repo name in query

* fix: extended query DSL for multiple types.

Synopsis output is now identical to master.

* feat: grabbing more metadata

Specifically names, owners, homepages of repos and orgs.

* feat: added stub for time filtering full tree.

* fix: added guards again null nodes.

Comes up when a repo doesn't have a master branch.

* feat: added filtering for full tree output.

* fix: grabbing review comments.

As well as their reactions if specified.

* fix: removed hanging test.

Testing that we need an auth token is unproductive and wasteful.

* Docs: lots of docstrings.

* docs: documented output formats in the README.

* testing: added tests for full tree output.