Daemonised (#64)
* feat: grabbing more metadata

Specifically names, owners, homepages of repos and orgs.

* fix: added guards again null nodes.

Comes up when a repo doesn't have a master branch.

* fix: pass errors through caching layer properly.

* fix: reject with useful failure info.

* feat: client now sleeps when limits are hit.

Added response handlers for abuse and rate limit exceeded errors which
cause the app to sleep for the amount of time in the `retry-after`
response header or until `x-ratelimit-reset` respectively. Nyc is now
capable of being left running indefinitely. If uninterrupted it will
eventually scrape anything.

Still does not daemonise or restart if interupted.

* feat: added flag to wipe cache before querying.

* fix: corrected ratelimit error catch.

Rate limit exceeded returns a 200 with a special body. Fun.

* fix: set request concurrency back to 1.

Was set higher for testing abuse alarm.

* feat: separated commits from commit comments.

Commit comments are a special GitHub feature and should always be
queried, whereas the commit log can be grabbed more efficiently in batch
operations by linking to a clone of the repo. The list of commit
comments is also in general much shorter than the commit log as well, so
query cost shouldn't be hit too hard. There is a cost, but this info is
important and there's nowhere to get it but the API.

* fix: bad merge.